Convert versions module to TS
