+++ package.json
@@ -30,6 +30,7 @@
   },
   "devDependencies": {
     "@types/node": "^12.0.10",
+    "@types/semver": "^6.0.1",
     "@typescript-eslint/eslint-plugin": "^1.11.0",
     "@typescript-eslint/parser": "^1.11.0",
     "eslint": "5.16.0",
